+++ main.py
@@ -1,0 +1,4 @@
+from src.web_launcher import launch_web_ui
+
+if __name__ == "__main__":
+    launch_web_ui()
