+++ src/web_api.py
@@ -1,30 +1,18 @@
 from src.global_console import GlobalConsole
-import webview
-from pathlib import Path
 from src.app_ui import AppUI
-import sys
-
-def web_refresh_targets_view():
-    if webview.windows:
-        webview.windows[0].evaluate_js("getTargets()")
-
-app_ui = AppUI(None, refresh_targets_view_fn=web_refresh_targets_view)
-
-def web_log_handler(msg):
-    if webview.windows:
-        webview.windows[0].evaluate_js(f"appendLog({repr(msg)})")
-
-GlobalConsole.set_web_handler(web_log_handler)
+import webview
 
 class WebApi:
+    def __init__(self, app_ui):
+        self.app_ui = app_ui
+
     def add_target(self):
-        result = app_ui.web_add_target_character()
+        result = self.app_ui.web_add_target_character()
         return result
-    
+
     def get_targets(self):
-        # Devuelve la lista de áreas monitoreadas con imagen en base64
         result = []
-        for area in app_ui.target_characters:
+        for area in self.app_ui.target_characters:
             img_b64 = AppUI.get_area_image_b64(area)
             result.append({
                 'start_x': area.start_x,
@@ -39,56 +27,39 @@ class WebApi:
 
     def update_target_pattern(self, index, new_pattern):
         try:
-            app_ui.target_characters[index].pattern_type = new_pattern
+            self.app_ui.target_characters[index].pattern_type = new_pattern
             GlobalConsole.log(f"Patrón del objetivo {index} actualizado a {new_pattern}")
             return True
         except Exception:
             return False
+
     def update_target_name(self, index, new_name):
         try:
-            old_name = app_ui.target_characters[index].name
+            old_name = self.app_ui.target_characters[index].name
             if old_name != new_name:
-                app_ui.target_characters[index].name = new_name
+                self.app_ui.target_characters[index].name = new_name
                 GlobalConsole.log(f"Nombre del objetivo {index} actualizado a {new_name}")
             return True
         except Exception:
             return False
 
     def toggle_monitoring(self):
-        app_ui.toggle_monitoring()
+        self.app_ui.toggle_monitoring()
 
     def delete_target(self, idx):
         try:
             self.last_target_character_will_be_deleted()
-            app_ui.delete_target_character(idx)
+            self.app_ui.delete_target_character(idx)
             return True
         except Exception:
             return False
-    
+
     def last_target_character_will_be_deleted(self):
-        if len(app_ui.target_characters) == 1 and app_ui.is_monitoring and webview.windows:
+        if len(self.app_ui.target_characters) == 1 and self.app_ui.is_monitoring and webview.windows:
             GlobalConsole.log("No quedan áreas para monitorear. Deteniendo el monitoreo.")
-            app_ui.toggle_monitoring()
+            self.app_ui.toggle_monitoring()
             webview.windows[0].evaluate_js("monitoringStatusChanged(false)")
 
     def close_app(self):
         if webview.windows:
             webview.windows[0].destroy()
-
-def launch_web_ui():
-    """Lanza la interfaz HTML usando pywebview en una ventana nueva."""
-    if hasattr(sys, '_MEIPASS'):
-        # PyInstaller: usar ruta temporal
-        html_path = Path(sys._MEIPASS) / 'src' / 'ui' / 'index.html'
-    else:
-        html_path = Path(__file__).parent / 'ui' / 'index.html'
-    webview.create_window('M2 Monitor',
-                          html_path.resolve().as_uri(),
-                          frameless=True,
-                          width=500,
-                          height=650,
-                          js_api=WebApi())
-    webview.start()
-
-if __name__ == "__main__":
-    launch_web_ui()
